Extract tests from test_local_storage into new files
